:recycle: :: 상태 변경 코드 분리
